Satellite_Cloud_Detection/requirements.txt
langchain-openai>=0.1.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
